[FIX] views.py & [DELETE] video_processing,py
